refactor(polls new): Separate form login into a new poll form component
