+++ src/polls/create/components/index.js
@@ -1,0 +1,3 @@
+import NewPollForm from './NewPollForm';
+
+export default NewPollForm;
